update Get_Distance to return meters rather than mm

+++ projects/01_obstacle_avoiding_robot/firmware/STM32_ObjectAvoider/Core/V3_CollisionAvoidance/Inc/main.h
@@ -121,7 +121,7 @@ void Command_Receive_Callback(void);
 void Update_State(void);
 
 // ultrasonic
-uint32_t Get_Distance(void);
+float Get_Distance(void);
 uint32_t micros(void);
 void delay_us(uint32_t micros);
